lecturer-review-backend/feedback/serializers.py
from rest_framework import serializers
from .models import Feedback

class FeedbackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Feedback
        fields = '__all__'  # or specify the fields you want to include, e.g., ['id', 'comment', 'rating']