lecturer-review-backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
django-filter
djangorestframework-simplejwt
django-environ